업데이트

+++ dev-community/dev-community-backend/src/main/java/com/letsgo/devcommunity/domain/member/dto/MemberProfileResponse.java
@@ -5,5 +5,5 @@ public record MemberProfileResponse(
         String profileImageUrl,
         int followerCount,
         int followingCount,
-        int receivedLikeCount
+        long receivedLikeCount
 ) { }
